proxy/requirements.txt
aiosmtpd==1.4.6
httpx==0.28.1
structlog==24.4.0
python-dateutil==2.9.0
